F_T30603 move MapBundle service to core
- adjust use statements
