+++ demosplan/DemosPlanCoreBundle/Logic/Map/ServiceStorage.php
@@ -13,8 +13,8 @@ namespace demosplan\DemosPlanCoreBundle\Logic\Map;
 use DemosEurope\DemosplanAddon\Contracts\Config\GlobalConfigInterface;
 use demosplan\DemosPlanCoreBundle\Exception\InvalidArgumentException;
 use demosplan\DemosPlanCoreBundle\Logic\LegacyFlashMessageCreator;
+use demosplan\DemosPlanCoreBundle\Services\Map\GetFeatureInfo;
 use demosplan\DemosPlanMapBundle\MapValidationException;
-use demosplan\DemosPlanMapBundle\Services\GetFeatureInfo\GetFeatureInfo;
 use Exception;
 use Psr\Log\LoggerInterface;
 use Symfony\Component\HttpKernel\Exception\HttpException;
